pyproject.toml
[project]
name = "tortoise-test"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)", "aerich (>=0.9.0,<0.10.0)"]

[[project.authors]]
name = "ulyana"
email = "ulyanaworkpage@gmail.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.3"

[tool.aerich]
tortoise_orm = "app.main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
